chore: better rename, fix unused variable warning (#99)
* Better refactor, fix unused var warn

* Better rename

* Remove xlayer suffix
